+++ index.html
@@ -60,6 +60,7 @@
       </section>
     </div>
   </div>
+  <div id="snackbar"></div>
 </body>
 
 </html>
