+++ src/css/app.css
@@ -97,29 +97,70 @@ body {
     color: gray;
 }
 
+#snackbar {
+    visibility: hidden;
+    min-width: 250px;
+    margin-left: -125px;
+    background-color: #333;
+    color: #fff;
+    text-align: center; 
+    border-radius: 2px;
+    padding: 16px;
+    position: fixed; 
+    z-index: 1; 
+    left: 50%; 
+    bottom: 30px; 
+  }
+  
+#snackbar.show {
+  visibility: visible; 
+  -webkit-animation: fadein 0.5s, fadeout 0.5s 2.5s;
+  animation: fadein 0.5s, fadeout 0.5s 2.5s;
+}
+  
+@-webkit-keyframes fadein {
+  from {bottom: 0; opacity: 0;}
+  to {bottom: 30px; opacity: 1;}
+}
+  
+@keyframes fadein {
+  from {bottom: 0; opacity: 0;}
+  to {bottom: 30px; opacity: 1;}
+}
+  
+@-webkit-keyframes fadeout {
+  from {bottom: 30px; opacity: 1;}
+  to {bottom: 0; opacity: 0;}
+}
+  
+@keyframes fadeout {
+  from {bottom: 30px; opacity: 1;}
+  to {bottom: 0; opacity: 0;}
+}
+
 @media screen and (max-width: 1280px) {
-    #app {
-        width: var(--large-list-size);
-    }
+  #app {
+      width: var(--large-list-size);
   }
+}
 
-  @media screen and (max-width: 960px) {
-    #app {
-        width: var(--medium-list-size);
-    }
+@media screen and (max-width: 960px) {
+  #app {
+      width: var(--medium-list-size);
   }
+}
 
-  @media screen and (max-width: 600px) {
-    #app {
-        width: var(--small-list-size);
-    }
-    #will-watch-button, #watched-button{
-        width: 90px;
-    }
-    .nav__button {
-        width: 40px;
-    }
-    .search-text{
-        display: none;
-    }
+@media screen and (max-width: 600px) {
+  #app {
+      width: var(--small-list-size);
   }
+  #will-watch-button, #watched-button{
+      width: 90px;
+  }
+  .nav__button {
+      width: 40px;
+  }
+  .search-text{
+      display: none;
+  }
+}
