+++ src/css/modal.css
@@ -232,8 +232,3 @@
         grid-template-columns: repeat(1, 1fr);
     }
   }
-
-
-
-  
-
