+++ src/js/Manager/SearchVideoManager.js
@@ -1,5 +1,5 @@
 import fetch from 'node-fetch';
-import { YOUTUBE_API_ENDPOINT } from '../youtubeApi';
+import { DUMMY_YOUTUBE_API_ENDPOINT } from '../youtubeApi';
 import { parseVideoInfo } from '../util';
 
 export default class SearchVideoManager {
@@ -29,7 +29,7 @@ export default class SearchVideoManager {
   }
 
   fetchYoutubeData(keyword) {
-    return fetch(YOUTUBE_API_ENDPOINT(keyword, this.nextPageToken)).then((response) => {
+    return fetch(DUMMY_YOUTUBE_API_ENDPOINT(keyword, this.nextPageToken)).then((response) => {
       if (!response.ok) {
         throw new Error(response.status);
       }
