+++ src/js/constants.js
@@ -1,5 +1,5 @@
 export const SEARCH_KEYWORD_MIN_LENGTH = 2;
-export const MAX_VIDEO_SAVE = 10;
+export const MAX_VIDEO_SAVE = 100;
 export const MAX_DATA_FETCH_AT_ONCE = 10;
 export const SCROLL_BUFFER_SECOND = 200;
 export const SCROLL_BUFFER_HEIGHT = 50;
@@ -12,6 +12,9 @@ export const ERROR_MESSAGE = {
 
 export const ALERT_MESSAGE = {
   NO_MORE_SEARCH_RESULT: '마지막 검색결과까지 모두 출력되었습니다.',
+  SAVED: '저장되었습니다',
+  CHANGED_TO_WILL_WATCH: '볼 영상으로 전환되었습니다',
+  CHANGED_TO_WATCHED: '본 영상으로 전환되었습니다',
 };
 
 export const REQUEST_PATH = 'youtube/v3/search';
